Add option to have raw observations in POMDP.step
We initially treated solely POMDP as a belief-space MDP env; now,
we add the flag POMDP.use_belif_space that dictates whether the
values returned from POMDP.reset and POMDP.step are beliefs or
raw observations.

To use Deep RL on raw observations, it is necessary to use RNN policies.
However, this makes the learning more challenging, and with the current
hyperparameters, Deep RL often fails to achieve good performance (over
10 seeds, only 3 got optimal policy in RedBlueAssistanceProblem). It is
possible that training for longer and with better hyperparameters would
improve the situation. In particular, tests with deep_rl_solve are
currently flaky.

Another thing: The POMDP class is getting more complicated as new
features are added; there might be a way to simplify it, so it might
be worth to spend some time thinking about this.