Fix black square images issue
rendering.Geom.__init__ adds a black color attribute, since it is a good default
for most geometries (e.g. lines, circles). However, for images, this causes the
whole image to be black, which is most often not the desired behavior.

So we set the color in rendering.Image.__init__ as white, which fixes
the issue.