Use random seed if no seed flag is passed
When the user does not use a seed flag, it seems that the best
behavior is to provide a random seed, instead of always using
the same one - I find myself having to change the seed flag
repeatedly just to see different behavior.

Instead of providing seed=None in this case, we select
a random seed, so that the results are also reproducible
(I can see the seed selected and reuse it if I want).